Update content to be pure AsciiDoc
When Pandoc converts Sphinx-Doc reStructuredText content to AsciiDoc, it
doesn't completely convert it to native AsciiDoc. This can cause
problems when attempting to create a PDF of the documentation (it breaks
the table of contents, among other things). So this commit converts the
remaining reStructuredText formatting to AsciiDoc-native syntax. It also
does a modest bit of text reflowing, so that it will follow the upcoming
writing style guide.